Fixed performance of `reader-node`
I tried to be clever and keep the file open and then sync when on block written was called. It was a bad idea and in fact cause bad performance because `onBlockWritten` is blocking and it seems it's much faster to write the file directly each time was well as serializing to `bytes` in memory.